Immobiliendatenbankanbindung richtig gemacht, löschen geht noch nicht

+++ Aufgabe 2 eigener Start/src/ersteVersuche/Immobilie/ImmobilienmenuWerkzeug.java
@@ -8,6 +8,7 @@ import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -17,6 +18,7 @@ import javax.swing.event.TableModelListener;
 import dis2011.DB2ConnectionManager;
 import ersteVersuche.Material.Haus;
 import ersteVersuche.Material.Immobilie;
+import ersteVersuche.Material.Makler;
 import ersteVersuche.Material.Wohnung;
 
 public class ImmobilienmenuWerkzeug {
@@ -56,13 +58,18 @@ public class ImmobilienmenuWerkzeug {
 				UpdImmobilie(i);
 			}
 		});
+		
+		for(Immobilie i :LadeImmobilien())
+		{
+			_GUI.GetTableModel().AddImmobilie(i);
+		}
+		
 	}
 
 	private void AddImmobilie() {
 		Immobilie i = _immobilieNeu.ErstelleImmobilie();
 
-		if (i != null )
-		{//&& AddImmobilieSQL(i)) {
+		if (i != null && AddImmobilieSQL(i)) {
 			_GUI.GetTableModel().AddImmobilie(i);
 		}
 		_GUI.repaint();
@@ -70,7 +77,7 @@ public class ImmobilienmenuWerkzeug {
 
 	private void DelImmobilie(Immobilie i) {
 
-		if (i != null){// && DelImmobilieSQL(i)) {
+		if (i != null && DelImmobilieSQL(i)) {
 			_GUI.GetTableModel().DeleteImmobilie(i);
 		}
 		_GUI.repaint();
@@ -78,7 +85,7 @@ public class ImmobilienmenuWerkzeug {
 
 	private void UpdImmobilie(Immobilie i) {
 
-		if (i != null){// && UpdImmobilieSQL(i)) {
+		if (i != null && UpdImmobilieSQL(i)) {
 		}
 		_GUI.repaint();
 	}
@@ -92,21 +99,51 @@ public class ImmobilienmenuWerkzeug {
 
 	private boolean AddImmobilieSQL(Immobilie i) {
 		Connection con = DB2ConnectionManager.getInstance().getConnection();
-		String insertSQL = "INSERT INTO Immobilie (ID, Ort, PLZ, Straße, Hausnummer, Fläche, Makler) VALUES(?, ?, ?, ?, ?, ?, ?, ?)";
+		String insertSQL = "INSERT INTO Immobilie ( Ort, PLZ, Straße, Hausnummer, Fläche, Makler) VALUES(?, ?, ?, ?, ?, ?)";
 		PreparedStatement pstmt;
 		try {
-			pstmt = con.prepareStatement(insertSQL);
-			pstmt.setInt(1, i.getID());
-			pstmt.setString(2, i.getOrt());
-			pstmt.setInt(3, i.getPLZ());
-			pstmt.setString(4, i.getStrasse());
-			pstmt.setInt(5, i.getHausNr());
-			pstmt.setFloat(6, i.getFlaeche());
-			pstmt.setString(7, Makler);
+			pstmt = con.prepareStatement(insertSQL,Statement.RETURN_GENERATED_KEYS);
+			pstmt.setString(1, i.getOrt());
+			pstmt.setInt(2, i.getPLZ());
+			pstmt.setString(3, i.getStrasse());
+			pstmt.setInt(4, i.getHausNr());
+			pstmt.setFloat(5, i.getFlaeche());
+			pstmt.setString(6, Makler);
 			// TODO: INSERT für immobilien
 			// Führe Anfrage aus
 			int rs = pstmt.executeUpdate();
-			return rs > 0;
+			if(rs <= 0) return false;
+			
+			ResultSet keys = pstmt.getGeneratedKeys();
+			keys.next();
+			int ID = keys.getInt(1);
+			if(i instanceof Haus)
+			{
+				insertSQL = "INSERT INTO Haus (ID, Stockwerke, Kaufpreis, Garten) VALUES(?, ?, ?, ?)";
+				pstmt = con.prepareStatement(insertSQL,Statement.RETURN_GENERATED_KEYS);
+				pstmt.setInt(1, ID);
+				pstmt.setInt(2, ((Haus) i).getStockwerke());
+				pstmt.setFloat(3, ((Haus) i).getKaufpreis());
+				pstmt.setBoolean(4, ((Haus) i).isGarten());
+				
+				rs = pstmt.executeUpdate();
+				if(rs > 0) return true;
+			}
+			if(i instanceof Wohnung)
+			{
+				insertSQL = "INSERT INTO Wohnung (ID, Stockwerk, Mietpreis, Zimmer, Balkon, EBK) VALUES(?, ?, ?, ?, ?, ?)";
+				pstmt = con.prepareStatement(insertSQL,Statement.RETURN_GENERATED_KEYS);
+				pstmt.setInt(1, ID);
+				pstmt.setInt(2, ((Wohnung) i).getStockwerk());
+				pstmt.setFloat(3, ((Wohnung) i).getMietpreis());
+				pstmt.setInt(4,((Wohnung) i).getZimmer());
+				pstmt.setBoolean(5, ((Wohnung) i).isBalkon());
+				pstmt.setBoolean(6, ((Wohnung) i).isEBK());
+				
+				rs = pstmt.executeUpdate();
+				if(rs > 0) return true;
+			}
+			
 		} catch (SQLException e) {
 
 			e.printStackTrace();
@@ -117,20 +154,29 @@ public class ImmobilienmenuWerkzeug {
 
 	private boolean DelImmobilieSQL(Immobilie i) {
 		Connection con = DB2ConnectionManager.getInstance().getConnection();
-		String selectSQL = "DELETE FROM Immobilie WHERE ID = ?; ";
+		String selectSQL ="";
 		if (i instanceof Haus)
 			selectSQL += " DELETE FROM Haus WHERE ID = ?";
 		if (i instanceof Wohnung)
 			selectSQL += " DELETE FROM Wohnung WHERE ID = ?";
+		
 		PreparedStatement pstmt;
 		try {
 			pstmt = con.prepareStatement(selectSQL);
 			pstmt.setInt(1, i.getID());
-			pstmt.setInt(2, i.getID());
+			//pstmt.setInt(2, i.getID());
 
 			// Führe Anfrage aus
 			int rs = pstmt.executeUpdate();
+			
+			//Und nun die Immobilie weg.
+			if(rs < 0) return false;
+			selectSQL = "DELETE FROM Immobilie WHERE ID = ? ";
+			pstmt = con.prepareStatement(selectSQL);
+			pstmt.setInt(1, i.getID());
+			rs = pstmt.executeUpdate();
 			return rs > 0;
+			
 		} catch (SQLException e) {
 
 			e.printStackTrace();
@@ -180,7 +226,7 @@ public class ImmobilienmenuWerkzeug {
 		Connection con = DB2ConnectionManager.getInstance().getConnection();
 
 		List<Haus> result = new ArrayList<Haus>();
-		String selectSQL = "SELECT * FROM Haus JOIN Immobilie ON Id";
+		String selectSQL = "SELECT * FROM Haus h, Immobilie i where h.id = i.id";
 		PreparedStatement pstmt;
 		try {
 			pstmt = con.prepareStatement(selectSQL);
@@ -190,8 +236,8 @@ public class ImmobilienmenuWerkzeug {
 
 			while (rs.next()) {
 				result.add(new Haus(rs.getInt("ID"), rs.getString("Ort"), rs
-						.getInt("PLZ"), rs.getString("Strasse"), rs
-						.getInt("HausNr"), rs.getFloat("Flaeche"), rs
+						.getInt("PLZ"), rs.getString("Straße"), rs
+						.getInt("Hausnummer"), rs.getFloat("Fläche"), rs
 						.getInt("Stockwerke"), rs.getFloat("Kaufpreis"), rs
 						.getBoolean("Garten")));
 			}
@@ -207,7 +253,7 @@ public class ImmobilienmenuWerkzeug {
 		Connection con = DB2ConnectionManager.getInstance().getConnection();
 
 		List<Wohnung> result = new ArrayList<Wohnung>();
-		String selectSQL = "SELECT * FROM Wohnung JOIN Immobilie ON Id";
+		String selectSQL = "SELECT * FROM Wohnung w, Immobilie i where w.id = i.id";
 		PreparedStatement pstmt;
 		try {
 			pstmt = con.prepareStatement(selectSQL);
@@ -217,8 +263,8 @@ public class ImmobilienmenuWerkzeug {
 
 			while (rs.next()) {
 				result.add(new Wohnung(rs.getInt("ID"), rs.getString("Ort"), rs
-						.getInt("PLZ"), rs.getString("Strasse"), Integer.parseInt(rs
-						.getString("HausNr")), rs.getFloat("Flaeche"), rs
+						.getInt("PLZ"), rs.getString("Straße"), Integer.parseInt(rs
+						.getString("Hausnummer")), rs.getFloat("Fläche"), rs
 						.getInt("Stockwerk"), rs.getFloat("Mietpreis"), rs
 						.getInt("Zimmer"), rs.getBoolean("Balkon"), rs
 						.getBoolean("EBK")));
